+++ Cargo.toml
@@ -188,7 +188,10 @@ databend-enterprise-storage-quota = { path = "src/query/ee_features/storage_quot
 databend-enterprise-stream-handler = { path = "src/query/ee_features/stream_handler" }
 databend-enterprise-vacuum-handler = { path = "src/query/ee_features/vacuum_handler" }
 databend-enterprise-virtual-column = { path = "src/query/ee_features/virtual_column" }
+databend-functions-scalar-arithmetic = { path = "src/query/functions/src/scalars/arithmetic" }
+databend-functions-scalar-arithmetic-modulo = { path = "src/query/functions/src/scalars/arithmetic_modulo" }
 databend-functions-scalar-datetime = { path = "src/query/functions/src/scalars/timestamp" }
+databend-functions-scalar-decimal-utils = { path = "src/query/functions/src/scalars/decimal" }
 databend-functions-scalar-geo = { path = "src/query/functions/src/scalars/geographic" }
 databend-functions-scalar-math = { path = "src/query/functions/src/scalars/mathematics" }
 databend-meta = { path = "src/meta/service" }
