+++ src/query/functions/src/scalars/arithmetic/Cargo.toml
@@ -1,15 +1,17 @@
 [package]
-name = "databend-functions-scalar-math"
+name = "databend-functions-scalar-arithmetic"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-crc32fast = { workspace = true }
-databend-common-base = { workspace = true }
 databend-common-expression = { workspace = true }
+databend-functions-scalar-arithmetic-modulo = { workspace = true }
 databend-functions-scalar-decimal-utils = { workspace = true }
-match-template = { workspace = true }
+ethnum = { workspace = true }
+lexical-core = { workspace = true }
 num-traits = { workspace = true }
+
+match-template = { workspace = true }
 
 [package.metadata.cargo-machete]
 ignored = ["match-template"]
