+++ src/query/functions/src/scalars/decimal/src/arithmetic.rs
