+++ src/query/functions/src/scalars/decimal/src/math.rs
