+++ src/query/functions/src/scalars/timestamp/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "databend-functions-scalar-datetime"
 version = "0.1.0"
-edition = "2024"
+edition = "2021"
 
 [dependencies]
 chrono = { workspace = true }
